LPS-74270 Routers should be in router package
