+++ modules/apps/headless-apio/organization/organization-apio-impl/src/main/java/com/liferay/organization/apio/internal/architect/router/PersonNestedCollectionRouter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.organization.apio.internal.architect.resource;
+package com.liferay.organization.apio.internal.architect.router;
 
 import com.liferay.apio.architect.pagination.PageItems;
 import com.liferay.apio.architect.pagination.Pagination;
